Drop testdata package
Initially, the idea was to define the configurations for all spki tests
in a central place: conf/testdata.

However, it is cleaner to define the configs close to the test, where
they are actually used.
