+++ go/tools/scion-pki/internal/v2/conf/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "as.go",
+        "doc.go",
         "issuer.go",
         "key.go",
         "trc.go",
@@ -32,10 +33,15 @@ go_test(
         "main_test.go",
         "trc_test.go",
     ],
-    data = ["//go/tools/scion-pki/internal/v2/conf/testdata:data"],
+    data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
-        "//go/tools/scion-pki/internal/v2/conf/testdata:go_default_library",
+        "//go/lib/addr:go_default_library",
+        "//go/lib/scrypto:go_default_library",
+        "//go/lib/scrypto/cert/v2:go_default_library",
+        "//go/lib/scrypto/trc/v2:go_default_library",
+        "//go/lib/util:go_default_library",
+        "//go/lib/xtest:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
