+++ go/tools/scion-pki/internal/v2/conf/as_test.go
@@ -18,12 +18,16 @@ import (
 	"bytes"
 	"io/ioutil"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/scionproto/scion/go/lib/addr"
+	"github.com/scionproto/scion/go/lib/scrypto"
+	"github.com/scionproto/scion/go/lib/util"
+	"github.com/scionproto/scion/go/lib/xtest"
 	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf"
-	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf/testdata"
 )
 
 func TestASEncode(t *testing.T) {
@@ -31,7 +35,7 @@ func TestASEncode(t *testing.T) {
 	require.NoError(t, err)
 
 	var buf bytes.Buffer
-	err = testdata.GoldenAS.Encode(&buf)
+	err = AS().Encode(&buf)
 	require.NoError(t, err)
 	assert.Equal(t, rawGolden, buf.Bytes())
 }
@@ -39,7 +43,7 @@ func TestASEncode(t *testing.T) {
 func TestLoadAS(t *testing.T) {
 	cfg, err := conf.LoadAS("testdata/as-v1.toml")
 	require.NoError(t, err)
-	assert.Equal(t, testdata.GoldenAS, cfg)
+	assert.Equal(t, AS(), cfg)
 }
 
 // TestUpdateGoldenAS provides an easy way to update the golden file after
@@ -47,9 +51,28 @@ func TestLoadAS(t *testing.T) {
 func TestUpdateGoldenAS(t *testing.T) {
 	if *update {
 		var buf bytes.Buffer
-		err := testdata.GoldenAS.Encode(&buf)
+		err := AS().Encode(&buf)
 		require.NoError(t, err)
 		err = ioutil.WriteFile("testdata/as-v1.toml", buf.Bytes(), 0644)
 		require.NoError(t, err)
+	}
+}
+
+// AS generates a AS certificate configuration for testing.
+func AS() conf.AS {
+	s, e, r := scrypto.KeyVersion(1), scrypto.KeyVersion(1), scrypto.KeyVersion(1)
+	return conf.AS{
+		Description:          "Testing AS",
+		Version:              1,
+		SigningKeyVersion:    &s,
+		EncryptionKeyVersion: &e,
+		RevocationKeyVersion: &r,
+		IssuerIA:             xtest.MustParseIA("1-ff00:0:110"),
+		IssuerCertVersion:    1,
+		OptDistPoints:        []addr.IA{xtest.MustParseIA("2-ff00:0:210")},
+		Validity: conf.Validity{
+			NotBefore: 42424242,
+			Validity:  util.DurWrap{Duration: 3 * 24 * time.Hour},
+		},
 	}
 }
