+++ go/tools/scion-pki/internal/v2/conf/issuer_test.go
@@ -18,12 +18,16 @@ import (
 	"bytes"
 	"io/ioutil"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/scionproto/scion/go/lib/addr"
+	"github.com/scionproto/scion/go/lib/scrypto"
+	"github.com/scionproto/scion/go/lib/util"
+	"github.com/scionproto/scion/go/lib/xtest"
 	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf"
-	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf/testdata"
 )
 
 func TestIssuerEncode(t *testing.T) {
@@ -31,7 +35,7 @@ func TestIssuerEncode(t *testing.T) {
 	require.NoError(t, err)
 
 	var buf bytes.Buffer
-	err = testdata.GoldenIssuer.Encode(&buf)
+	err = Issuer().Encode(&buf)
 	require.NoError(t, err)
 	assert.Equal(t, rawGolden, buf.Bytes())
 
@@ -40,7 +44,7 @@ func TestIssuerEncode(t *testing.T) {
 func TestLoadIssuer(t *testing.T) {
 	cfg, err := conf.LoadIssuer("testdata/issuer-v1.toml")
 	require.NoError(t, err)
-	assert.Equal(t, testdata.GoldenIssuer, cfg)
+	assert.Equal(t, Issuer(), cfg)
 }
 
 // TestUpdateGoldenIssuer provides an easy way to update the golden file after
@@ -48,9 +52,26 @@ func TestLoadIssuer(t *testing.T) {
 func TestUpdateGoldenIssuer(t *testing.T) {
 	if *update {
 		var buf bytes.Buffer
-		err := testdata.GoldenIssuer.Encode(&buf)
+		err := Issuer().Encode(&buf)
 		require.NoError(t, err)
 		err = ioutil.WriteFile("testdata/issuer-v1.toml", buf.Bytes(), 0644)
 		require.NoError(t, err)
+	}
+}
+
+// Issuer generates a issuer certificate configuration for testing.
+func Issuer() conf.Issuer {
+	i, r := scrypto.KeyVersion(1), scrypto.KeyVersion(1)
+	return conf.Issuer{
+		Description:          "Testing Issuer",
+		Version:              1,
+		IssuingKeyVersion:    &i,
+		RevocationKeyVersion: &r,
+		TRCVersion:           1,
+		OptDistPoints:        []addr.IA{xtest.MustParseIA("2-ff00:0:210")},
+		Validity: conf.Validity{
+			NotBefore: 42424242,
+			Validity:  util.DurWrap{Duration: 3 * 24 * time.Hour},
+		},
 	}
 }
