+++ go/tools/scion-pki/internal/v2/conf/key_test.go
@@ -18,12 +18,16 @@ import (
 	"bytes"
 	"io/ioutil"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/scionproto/scion/go/lib/scrypto"
+	"github.com/scionproto/scion/go/lib/scrypto/cert/v2"
+	"github.com/scionproto/scion/go/lib/scrypto/trc/v2"
+	"github.com/scionproto/scion/go/lib/util"
 	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf"
-	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf/testdata"
 )
 
 func TestKeysEncode(t *testing.T) {
@@ -31,7 +35,7 @@ func TestKeysEncode(t *testing.T) {
 	require.NoError(t, err)
 
 	var buf bytes.Buffer
-	err = testdata.GoldenKeys.Encode(&buf)
+	err = Keys().Encode(&buf)
 	require.NoError(t, err)
 	assert.Equal(t, rawGolden, buf.Bytes())
 }
@@ -39,7 +43,7 @@ func TestKeysEncode(t *testing.T) {
 func TestLoadKeys(t *testing.T) {
 	keys, err := conf.LoadKeys("testdata/keys.toml")
 	require.NoError(t, err)
-	assert.Equal(t, testdata.GoldenKeys, keys)
+	assert.Equal(t, Keys(), keys)
 }
 
 // TestUpdateGoldenKeys provides an easy way to update the golden file after
@@ -47,9 +51,56 @@ func TestLoadKeys(t *testing.T) {
 func TestUpdateGoldenKeys(t *testing.T) {
 	if *update {
 		var buf bytes.Buffer
-		err := testdata.GoldenKeys.Encode(&buf)
+		err := Keys().Encode(&buf)
 		require.NoError(t, err)
 		err = ioutil.WriteFile("testdata/keys.toml", buf.Bytes(), 0644)
 		require.NoError(t, err)
+	}
+}
+
+// Keys generates a key configuration for testing.
+func Keys() conf.Keys {
+	return conf.Keys{
+		Primary: map[trc.KeyType]map[scrypto.KeyVersion]conf.KeyMeta{
+			trc.IssuingKey: {
+				1: keyMeta(scrypto.Ed25519, 42424242, 365*24*time.Hour),
+			},
+			trc.OfflineKey: {
+				1: keyMeta(scrypto.Ed25519, 42424242, 365*24*time.Hour),
+			},
+			trc.OnlineKey: {
+				1: keyMeta(scrypto.Ed25519, 42424242, 365*24*time.Hour),
+				2: keyMeta(scrypto.Ed25519, 42424242, 365*24*time.Hour),
+			},
+		},
+		Issuer: map[cert.KeyType]map[scrypto.KeyVersion]conf.KeyMeta{
+			cert.IssuingKey: {
+				1: keyMeta(scrypto.Ed25519, 42424242, 180*24*time.Hour),
+			},
+			cert.RevocationKey: {
+				2: keyMeta(scrypto.Ed25519, 42424242, 180*24*time.Hour),
+			},
+		},
+		AS: map[cert.KeyType]map[scrypto.KeyVersion]conf.KeyMeta{
+			cert.EncryptionKey: {
+				1: keyMeta(scrypto.Curve25519xSalsa20Poly1305, 42424242, 90*24*time.Hour),
+			},
+			cert.RevocationKey: {
+				2: keyMeta(scrypto.Ed25519, 42424242, 90*24*time.Hour),
+			},
+			cert.SigningKey: {
+				3: keyMeta(scrypto.Ed25519, 42424242, 90*24*time.Hour),
+			},
+		},
+	}
+}
+
+func keyMeta(algo string, notBefore uint32, validity time.Duration) conf.KeyMeta {
+	return conf.KeyMeta{
+		Algorithm: algo,
+		Validity: conf.Validity{
+			NotBefore: notBefore,
+			Validity:  util.DurWrap{Duration: validity},
+		},
 	}
 }
