--- go/tools/scion-pki/internal/v2/conf/testdata/BUILD.bazel
@@ -1,28 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-filegroup(
-    name = "data",
-    srcs = glob(["*.toml"]),
-    visibility = ["//visibility:public"],
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "as.go",
-        "issuer.go",
-        "keys.go",
-        "trc.go",
-    ],
-    importpath = "github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf/testdata",
-    visibility = ["//go/tools/scion-pki:__subpackages__"],
-    deps = [
-        "//go/lib/addr:go_default_library",
-        "//go/lib/scrypto:go_default_library",
-        "//go/lib/scrypto/cert/v2:go_default_library",
-        "//go/lib/scrypto/trc/v2:go_default_library",
-        "//go/lib/util:go_default_library",
-        "//go/lib/xtest:go_default_library",
-        "//go/tools/scion-pki/internal/v2/conf:go_default_library",
-    ],
-)
