+++ go/tools/scion-pki/internal/v2/conf/trc_test.go
@@ -18,12 +18,17 @@ import (
 	"bytes"
 	"io/ioutil"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/scionproto/scion/go/lib/addr"
+	"github.com/scionproto/scion/go/lib/scrypto"
+	"github.com/scionproto/scion/go/lib/scrypto/trc/v2"
+	"github.com/scionproto/scion/go/lib/util"
+	"github.com/scionproto/scion/go/lib/xtest"
 	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf"
-	"github.com/scionproto/scion/go/tools/scion-pki/internal/v2/conf/testdata"
 )
 
 func TestTRCEncode(t *testing.T) {
@@ -33,11 +38,11 @@ func TestTRCEncode(t *testing.T) {
 	}{
 		"v1": {
 			File:   "testdata/trc-v1.toml",
-			Config: testdata.GoldenTRCv1,
+			Config: TRCv1(),
 		},
 		"v2": {
 			File:   "testdata/trc-v2.toml",
-			Config: testdata.GoldenTRCv2,
+			Config: TRCv2(),
 		},
 	}
 	for name, test := range tests {
@@ -60,11 +65,11 @@ func TestLoadTRC(t *testing.T) {
 	}{
 		"v1": {
 			File:   "testdata/trc-v1.toml",
-			Config: testdata.GoldenTRCv1,
+			Config: TRCv1(),
 		},
 		"v2": {
 			File:   "testdata/trc-v2.toml",
-			Config: testdata.GoldenTRCv2,
+			Config: TRCv2(),
 		},
 	}
 	for name, test := range tests {
@@ -81,8 +86,8 @@ func TestLoadTRC(t *testing.T) {
 func TestUpdateGoldenTRC(t *testing.T) {
 	if *update {
 		cfgs := map[string]conf.TRC2{
-			"testdata/trc-v1.toml": testdata.GoldenTRCv1,
-			"testdata/trc-v2.toml": testdata.GoldenTRCv2,
+			"testdata/trc-v1.toml": TRCv1(),
+			"testdata/trc-v2.toml": TRCv2(),
 		}
 		for file, cfg := range cfgs {
 			var buf bytes.Buffer
@@ -91,5 +96,62 @@ func TestUpdateGoldenTRC(t *testing.T) {
 			err = ioutil.WriteFile(file, buf.Bytes(), 0644)
 			require.NoError(t, err)
 		}
+	}
+}
+
+// TRCv1 generates a TRC configuration for testing.
+func TRCv1() conf.TRC2 {
+	t := true
+	v1 := scrypto.KeyVersion(1)
+	return conf.TRC2{
+		Description:       "Testing TRC",
+		Version:           1,
+		BaseVersion:       1,
+		VotingQuorum:      1,
+		GracePeriod:       util.DurWrap{},
+		TrustResetAllowed: &t,
+		Votes:             []addr.AS{},
+		Validity: conf.Validity{
+			NotBefore: 42424242,
+			Validity:  util.DurWrap{Duration: 5 * 24 * time.Hour},
+		},
+		PrimaryASes: map[addr.AS]conf.Primary{
+			xtest.MustParseAS("ff00:0:110"): {
+				Attributes: []trc.Attribute{trc.Authoritative, trc.Core,
+					trc.Issuing, trc.Voting},
+				IssuingKeyVersion:       &v1,
+				VotingOnlineKeyVersion:  &v1,
+				VotingOfflineKeyVersion: &v1,
+			},
+		},
+	}
+}
+
+// TRCv2 generates a TRC configuration for testing.
+func TRCv2() conf.TRC2 {
+	t := true
+	v1 := scrypto.KeyVersion(1)
+	v2 := scrypto.KeyVersion(2)
+	return conf.TRC2{
+		Description:       "Testing TRC",
+		Version:           2,
+		BaseVersion:       1,
+		VotingQuorum:      1,
+		GracePeriod:       util.DurWrap{Duration: time.Hour},
+		TrustResetAllowed: &t,
+		Votes:             []addr.AS{xtest.MustParseAS("ff00:0:110")},
+		Validity: conf.Validity{
+			NotBefore: 42424248,
+			Validity:  util.DurWrap{Duration: 5 * 24 * time.Hour},
+		},
+		PrimaryASes: map[addr.AS]conf.Primary{
+			xtest.MustParseAS("ff00:0:110"): {
+				Attributes: []trc.Attribute{trc.Authoritative, trc.Core,
+					trc.Issuing, trc.Voting},
+				IssuingKeyVersion:       &v1,
+				VotingOnlineKeyVersion:  &v2,
+				VotingOfflineKeyVersion: &v1,
+			},
+		},
 	}
 }
